+++ smart-contracts/test/Lock/hooks/erc1155BalanceOfHook.js
@@ -1,18 +1,20 @@
 const { constants } = require('hardlydifficult-ethereum-contracts')
 const { reverts } = require('truffle-assertions')
-const deployLocks = require('../helpers/deployLocks')
+const deployLocks = require('../../helpers/deployLocks')
 
 const unlockContract = artifacts.require('Unlock.sol')
-const Erc721TokenUriHook = artifacts.require('ERC721BalanceOfHook')
-const TestERC721 = artifacts.require('TestERC721')
-const getProxy = require('../helpers/proxy')
+const Erc1155TokenUriHook = artifacts.require('ERC1155BalanceOfHook')
+const TestERC1155 = artifacts.require('TestERC1155')
+const getProxy = require('../../helpers/proxy')
 
 let lock
 let unlock
 let hook
 let nft
 
-contract('ERC721BalanceOfHook', (accounts) => {
+const GOLD = 1
+
+contract('ERC1155BalanceOfHook', (accounts) => {
   const from = accounts[1]
   const nftOwner = accounts[2]
   const keyOwner = accounts[3]
@@ -22,11 +24,11 @@ contract('ERC721BalanceOfHook', (accounts) => {
     const locks = await deployLocks(unlock, accounts[0])
     lock = locks.FIRST
 
-    // deploy some ERC721
-    nft = await TestERC721.new()
+    // deploy some ERC1155
+    nft = await TestERC1155.new()
 
     // deploy the hook
-    hook = await Erc721TokenUriHook.new()
+    hook = await Erc1155TokenUriHook.new()
 
     // set the hook
     await lock.setEventHooks(
@@ -39,25 +41,33 @@ contract('ERC721BalanceOfHook', (accounts) => {
 
   describe('setting mapping', () => {
     beforeEach(async () => {
-      await hook.createMapping(lock.address, nft.address)
+      await hook.createMapping(lock.address, nft.address, GOLD)
     })
-    it('should record the corresponding erc721 address', async () => {
+
+    it('should record the corresponding NFT address', async () => {
       assert.equal(await hook.nftAddresses(lock.address), nft.address)
     })
+
+    it('should record the corresponding token type', async () => {
+      assert.equal(await hook.nftTokenIds(lock.address), GOLD)
+    })
+
     it('should only allow lock managers to set mapping', async () => {
       await reverts(
-        hook.createMapping(lock.address, nft.address, { from: accounts[5] }),
+        hook.createMapping(lock.address, nft.address, GOLD, {
+          from: accounts[5],
+        }),
         'Caller does not have the LockManager role'
       )
     })
     it('throws on zero addresses', async () => {
       await reverts(
-        hook.createMapping(constants.ZERO_ADDRESS, nft.address),
+        hook.createMapping(constants.ZERO_ADDRESS, nft.address, GOLD),
         'Lock address can not be zero'
       )
       await reverts(
-        hook.createMapping(lock.address, constants.ZERO_ADDRESS),
-        'ERC721 address can not be zero'
+        hook.createMapping(lock.address, constants.ZERO_ADDRESS, GOLD),
+        'ERC1155 address can not be zero'
       )
     })
   })
@@ -74,7 +84,7 @@ contract('ERC721BalanceOfHook', (accounts) => {
         [keyOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [[]],
+        [],
         {
           from,
           value: keyPrice,
@@ -85,22 +95,21 @@ contract('ERC721BalanceOfHook', (accounts) => {
     it('with an expired key', async () => {
       // buy a key
       const keyPrice = await lock.keyPrice()
-      const tx = await lock.purchase(
+      await lock.purchase(
         [],
         [keyOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [[]],
+        [],
         {
           from,
           value: keyPrice,
         }
       )
-      const { args } = tx.logs.find((v) => v.event === 'Transfer')
       assert.equal(await lock.getHasValidKey(keyOwner), true)
 
       // expire the key
-      await lock.expireAndRefundFor(args.tokenId, 0)
+      await lock.expireAndRefundFor(keyOwner, 0)
       assert.equal(await lock.getHasValidKey(keyOwner), false)
     })
   })
@@ -108,9 +117,9 @@ contract('ERC721BalanceOfHook', (accounts) => {
   describe('mapping is set, account holds a nft', () => {
     beforeEach(async () => {
       // mint one token
-      await nft.mint(nftOwner)
+      await nft.mint(nftOwner, GOLD)
       // create mapping
-      await hook.createMapping(lock.address, nft.address)
+      await hook.createMapping(lock.address, nft.address, GOLD)
     })
 
     it('with no valid key', async () => {
@@ -124,7 +133,7 @@ contract('ERC721BalanceOfHook', (accounts) => {
         [nftOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [[]],
+        [],
         {
           from,
           value: keyPrice,
@@ -135,12 +144,12 @@ contract('ERC721BalanceOfHook', (accounts) => {
     it('with an expired key', async () => {
       // buy a key
       const keyPrice = await lock.keyPrice()
-      const tx = await lock.purchase(
+      await lock.purchase(
         [],
         [nftOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [[]],
+        [],
         {
           from,
           value: keyPrice,
@@ -149,8 +158,7 @@ contract('ERC721BalanceOfHook', (accounts) => {
       assert.equal(await lock.getHasValidKey(nftOwner), true)
 
       // expire the key
-      const { args } = tx.logs.find((v) => v.event === 'Transfer')
-      await lock.expireAndRefundFor(args.tokenId, 0)
+      await lock.expireAndRefundFor(nftOwner, 0)
       assert.equal(await lock.getHasValidKey(nftOwner), true)
     })
   })
@@ -167,7 +175,7 @@ contract('ERC721BalanceOfHook', (accounts) => {
         [keyOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [[]],
+        [],
         {
           from,
           value: keyPrice,
@@ -178,12 +186,12 @@ contract('ERC721BalanceOfHook', (accounts) => {
     it('with an expired key', async () => {
       // buy a key
       const keyPrice = await lock.keyPrice()
-      const tx = await lock.purchase(
+      await lock.purchase(
         [],
         [keyOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [[]],
+        [],
         {
           from,
           value: keyPrice,
@@ -192,8 +200,7 @@ contract('ERC721BalanceOfHook', (accounts) => {
       assert.equal(await lock.getHasValidKey(keyOwner), true)
 
       // expire the key
-      const { args } = tx.logs.find((v) => v.event === 'Transfer')
-      await lock.expireAndRefundFor(args.tokenId, 0)
+      await lock.expireAndRefundFor(keyOwner, 0)
       assert.equal(await lock.getHasValidKey(keyOwner), false)
     })
   })
