+++ smart-contracts/test/Lock/hooks/erc20BalanceOfHook.js
@@ -1,12 +1,12 @@
 const { ethers } = require('hardhat')
 const { constants } = require('hardlydifficult-ethereum-contracts')
 const { reverts } = require('truffle-assertions')
-const deployLocks = require('../helpers/deployLocks')
+const deployLocks = require('../../helpers/deployLocks')
 
 const unlockContract = artifacts.require('Unlock.sol')
 const Erc20TokenUriHook = artifacts.require('ERC20BalanceOfHook')
 const TestERC20 = artifacts.require('TestERC20')
-const getProxy = require('../helpers/proxy')
+const getProxy = require('../../helpers/proxy')
 
 let lock
 let unlock
