+++ smart-contracts/test/Lock/hooks/erc721BalanceOfHook.js
@@ -3,18 +3,16 @@ const { reverts } = require('truffle-assertions')
 const deployLocks = require('../helpers/deployLocks')
 
 const unlockContract = artifacts.require('Unlock.sol')
-const Erc1155TokenUriHook = artifacts.require('ERC1155BalanceOfHook')
-const TestERC1155 = artifacts.require('TestERC1155')
-const getProxy = require('../helpers/proxy')
+const Erc721TokenUriHook = artifacts.require('ERC721BalanceOfHook')
+const TestERC721 = artifacts.require('TestERC721')
+const getProxy = require('../../helpers/proxy')
 
 let lock
 let unlock
 let hook
 let nft
 
-const GOLD = 1
-
-contract('ERC1155BalanceOfHook', (accounts) => {
+contract('ERC721BalanceOfHook', (accounts) => {
   const from = accounts[1]
   const nftOwner = accounts[2]
   const keyOwner = accounts[3]
@@ -24,11 +22,11 @@ contract('ERC1155BalanceOfHook', (accounts) => {
     const locks = await deployLocks(unlock, accounts[0])
     lock = locks.FIRST
 
-    // deploy some ERC1155
-    nft = await TestERC1155.new()
+    // deploy some ERC721
+    nft = await TestERC721.new()
 
     // deploy the hook
-    hook = await Erc1155TokenUriHook.new()
+    hook = await Erc721TokenUriHook.new()
 
     // set the hook
     await lock.setEventHooks(
@@ -41,33 +39,25 @@ contract('ERC1155BalanceOfHook', (accounts) => {
 
   describe('setting mapping', () => {
     beforeEach(async () => {
-      await hook.createMapping(lock.address, nft.address, GOLD)
+      await hook.createMapping(lock.address, nft.address)
     })
-
-    it('should record the corresponding NFT address', async () => {
+    it('should record the corresponding erc721 address', async () => {
       assert.equal(await hook.nftAddresses(lock.address), nft.address)
     })
-
-    it('should record the corresponding token type', async () => {
-      assert.equal(await hook.nftTokenIds(lock.address), GOLD)
-    })
-
     it('should only allow lock managers to set mapping', async () => {
       await reverts(
-        hook.createMapping(lock.address, nft.address, GOLD, {
-          from: accounts[5],
-        }),
+        hook.createMapping(lock.address, nft.address, { from: accounts[5] }),
         'Caller does not have the LockManager role'
       )
     })
     it('throws on zero addresses', async () => {
       await reverts(
-        hook.createMapping(constants.ZERO_ADDRESS, nft.address, GOLD),
+        hook.createMapping(constants.ZERO_ADDRESS, nft.address),
         'Lock address can not be zero'
       )
       await reverts(
-        hook.createMapping(lock.address, constants.ZERO_ADDRESS, GOLD),
-        'ERC1155 address can not be zero'
+        hook.createMapping(lock.address, constants.ZERO_ADDRESS),
+        'ERC721 address can not be zero'
       )
     })
   })
@@ -84,7 +74,7 @@ contract('ERC1155BalanceOfHook', (accounts) => {
         [keyOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [],
+        [[]],
         {
           from,
           value: keyPrice,
@@ -95,21 +85,22 @@ contract('ERC1155BalanceOfHook', (accounts) => {
     it('with an expired key', async () => {
       // buy a key
       const keyPrice = await lock.keyPrice()
-      await lock.purchase(
+      const tx = await lock.purchase(
         [],
         [keyOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [],
+        [[]],
         {
           from,
           value: keyPrice,
         }
       )
+      const { args } = tx.logs.find((v) => v.event === 'Transfer')
       assert.equal(await lock.getHasValidKey(keyOwner), true)
 
       // expire the key
-      await lock.expireAndRefundFor(keyOwner, 0)
+      await lock.expireAndRefundFor(args.tokenId, 0)
       assert.equal(await lock.getHasValidKey(keyOwner), false)
     })
   })
@@ -117,9 +108,9 @@ contract('ERC1155BalanceOfHook', (accounts) => {
   describe('mapping is set, account holds a nft', () => {
     beforeEach(async () => {
       // mint one token
-      await nft.mint(nftOwner, GOLD)
+      await nft.mint(nftOwner)
       // create mapping
-      await hook.createMapping(lock.address, nft.address, GOLD)
+      await hook.createMapping(lock.address, nft.address)
     })
 
     it('with no valid key', async () => {
@@ -133,7 +124,7 @@ contract('ERC1155BalanceOfHook', (accounts) => {
         [nftOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [],
+        [[]],
         {
           from,
           value: keyPrice,
@@ -144,12 +135,12 @@ contract('ERC1155BalanceOfHook', (accounts) => {
     it('with an expired key', async () => {
       // buy a key
       const keyPrice = await lock.keyPrice()
-      await lock.purchase(
+      const tx = await lock.purchase(
         [],
         [nftOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [],
+        [[]],
         {
           from,
           value: keyPrice,
@@ -158,7 +149,8 @@ contract('ERC1155BalanceOfHook', (accounts) => {
       assert.equal(await lock.getHasValidKey(nftOwner), true)
 
       // expire the key
-      await lock.expireAndRefundFor(nftOwner, 0)
+      const { args } = tx.logs.find((v) => v.event === 'Transfer')
+      await lock.expireAndRefundFor(args.tokenId, 0)
       assert.equal(await lock.getHasValidKey(nftOwner), true)
     })
   })
@@ -175,7 +167,7 @@ contract('ERC1155BalanceOfHook', (accounts) => {
         [keyOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [],
+        [[]],
         {
           from,
           value: keyPrice,
@@ -186,12 +178,12 @@ contract('ERC1155BalanceOfHook', (accounts) => {
     it('with an expired key', async () => {
       // buy a key
       const keyPrice = await lock.keyPrice()
-      await lock.purchase(
+      const tx = await lock.purchase(
         [],
         [keyOwner],
         [constants.ZERO_ADDRESS],
         [constants.ZERO_ADDRESS],
-        [],
+        [[]],
         {
           from,
           value: keyPrice,
@@ -200,7 +192,8 @@ contract('ERC1155BalanceOfHook', (accounts) => {
       assert.equal(await lock.getHasValidKey(keyOwner), true)
 
       // expire the key
-      await lock.expireAndRefundFor(keyOwner, 0)
+      const { args } = tx.logs.find((v) => v.event === 'Transfer')
+      await lock.expireAndRefundFor(args.tokenId, 0)
       assert.equal(await lock.getHasValidKey(keyOwner), false)
     })
   })
